# Rewrite & optimize LimitedQueue