--- a/src/common/Channel.cpp
+++ b/src/common/Channel.cpp
@@ -246,23 +246,16 @@ void Channel::deleteMessage(QString messageID)
         msg->flags.set(MessageFlag::Disabled);
     }
 }
+
 MessagePtr Channel::findMessage(QString messageID)
 {
-    LimitedQueueSnapshot<MessagePtr> snapshot = this->getMessageSnapshot();
-    int snapshotLength = snapshot.size();
+    MessagePtr res;
 
-    int end = std::max(0, snapshotLength - 200);
+    this->messages_.rfind(res, [&messageID](const MessagePtr &msg) {
+        return msg->id == messageID;
+    });
 
-    for (int i = snapshotLength - 1; i >= end; --i)
-    {
-        auto &s = snapshot[i];
-
-        if (s->id == messageID)
-        {
-            return s;
-        }
-    }
-    return nullptr;
+    return res;
 }
 
 bool Channel::canSendMessage() const