--- a/src/messages/LimitedQueue.hpp
+++ b/src/messages/LimitedQueue.hpp
@@ -3,6 +3,8 @@
 #include "messages/LimitedQueueSnapshot.hpp"
 
 #include <boost/circular_buffer.hpp>
+
+#include <cassert>
 #include <mutex>
 #include <vector>
 
@@ -17,6 +19,8 @@ public:
         , buffer_(limit)
     {
     }
+
+    // Property Accessors
 
     size_t size() const
     {
@@ -38,6 +42,38 @@ public:
         return this->buffer_.full();
     }
 
+    size_t space() const
+    {
+        return this->limit() - this->size();
+    }
+
+    // Value Accessors
+    // copies of values are returned so that references aren't invalidated
+
+    T at(size_t index) const
+    {
+        std::lock_guard<std::mutex> lock(this->mutex_);
+
+        assert(index < this->buffer_.size());
+        return this->buffer_[index];
+    }
+
+    T front() const
+    {
+        std::lock_guard<std::mutex> lock(this->mutex_);
+
+        return this->buffer_.front();
+    }
+
+    T back() const
+    {
+        std::lock_guard<std::mutex> lock(this->mutex_);
+
+        return this->buffer_.back();
+    }
+
+    // Modifiers
+
     void clear()
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
@@ -45,6 +81,8 @@ public:
         this->buffer_.clear();
     }
 
+    // Pushes an item to the end of the queue. If an element is removed from
+    // the front, true is returned and deleted is set.
     bool pushBack(const T &item, T &deleted)
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
@@ -58,41 +96,39 @@ public:
         return full;
     }
 
+    // Pushes an item to the end of the queue.
     bool pushBack(const T &item)
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
 
         bool full = this->buffer_.full();
-        if (full)
-        {
-            this->buffer_.front();
-        }
         this->buffer_.push_back(item);
         return full;
     }
 
+    // Pushes as many items as possible from the end of the given vector, until
+    // the queue is full. Returns the subset of items that was pushed.
     std::vector<T> pushFront(const std::vector<T> &items)
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
 
+        size_t numToPush = std::min(items.size(), this->space());
         std::vector<T> pushed;
-        pushed.reserve(
-            std::min(items.size(), this->limit_ - this->buffer_.size()));
+        pushed.reserve(numToPush);
 
-        for (auto it = items.crbegin(); it != items.crend(); ++it)
+        size_t f = items.size() - numToPush;
+        size_t b = items.size() - 1;
+        for (; f < items.size(); ++f, --b)
         {
-            if (this->buffer_.full())
-            {
-                break;
-            }
-
-            this->buffer_.push_front(*it);
-            pushed.push_back(*it);
+            this->buffer_.push_front(items[b]);
+            pushed.push_back(items[f]);
         }
 
         return pushed;
     }
 
+    // Replaces the given item with a replacement. Returns the index of the
+    // replacement, or -1 if the item was not found.
     int replaceItem(const T &item, const T &replacement)
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
@@ -108,6 +144,8 @@ public:
         return -1;
     }
 
+    // Attempts to replace the item at the given index. Returns whether the
+    // replacement succeeded.
     bool replaceItem(size_t index, const T &replacement)
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
@@ -125,6 +163,45 @@ public:
     {
         std::lock_guard<std::mutex> lock(this->mutex_);
         return LimitedQueueSnapshot<T>(this->buffer_);
+    }
+
+    // Actions
+
+    // Finds the first item that matches the given predicate. If an item is
+    // found, result is set and true is returned.
+    template <typename Predicate>
+    bool find(T &result, Predicate pred) const
+    {
+        std::lock_guard<std::mutex> lock(this->mutex_);
+
+        for (const auto &item : this->buffer_)
+        {
+            if (pred(item))
+            {
+                result = item;
+                return true;
+            }
+        }
+        return false;
+    }
+
+    // Finds the first item that matches the given predicate, starting at the
+    // end and working towards the beginning. If an item is found, result is
+    // set and true is returned.
+    template <typename Predicate>
+    bool rfind(T &result, Predicate pred) const
+    {
+        std::lock_guard<std::mutex> lock(this->mutex_);
+
+        for (auto it = this->buffer_.rbegin(); it != this->buffer_.rend(); ++it)
+        {
+            if (pred(*it))
+            {
+                result = *it;
+                return true;
+            }
+        }
+        return false;
     }
 
 private: