… CHANGELOG.md
… benchmarks/CMakeLists.txt
… benchmarks/src/LimitedQueue.cpp
… src/messages/LimitedQueueSnapshot.hpp
… src/widgets/Scrollbar.cpp
… src/widgets/Scrollbar.hpp
… src/widgets/helper/ChannelView.cpp
… src/widgets/helper/ChannelView.hpp
… tests/CMakeLists.txt
… tests/src/LimitedQueue.cpp