# Description
Rewrites `LimitedQueue<T>` to use a circular buffer instead of allocating and moving around tons of `shared_ptrs`. Adds some additional features to the limited queue like `find`/`rfind` so that taking snapshots of the entire queue is less necessary. `LimitedQueueSnapshot<T>` has also been updated with convenience functions like `begin`/`end` for use in range-based for loops/libraries.

Attempting to profile the application on macOS with Instruments (lol) shows a minor speedup and a bit lower memory usage, but I'm not sure if I'm really using the profiler correctly. I added some benchmarks and tests, too.

This rewrite also seems to fix #1421, where Chatterino would crash on receiving a message after loading 1000 historical messages.

**Disclaimer: I was not 100% sure which methods in `LimitedQueue` did _not_ need to be guarded by the mutex, so please verify my decisions.**

<details>
<summary>Benchmark (w/ mutex locking)</summary>

This PR (5af89c3):
<pre>
Run on (8 X 24.1208 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x8)
Load Average: 4.86, 3.55, 2.54
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_LimitedQueue_PushBack                     19.4 ns         19.4 ns     34152006
BM_LimitedQueue_PushFront_One                 670 ns          675 ns      1039594
BM_LimitedQueue_PushFront_Many               2834 ns         2834 ns       247147
BM_LimitedQueue_Replace                       261 ns          261 ns      2678934
BM_LimitedQueue_Snapshot                     1030 ns         1030 ns       679704
BM_LimitedQueue_Snapshot_ExpensiveCopy       3710 ns         3710 ns       188782
BM_LimitedQueue_Find                         1729 ns         1729 ns       405879
</pre>
</details>

## Pull request checklist:

- [x] `CHANGELOG.md` was updated, if applicable