@iProdigy 2022-06-06
Agree, can resolve the shared lock concerns if the methods were private (as the calling scope likely would've already acquired a lock)