@dnsge 2022-06-07
I added most of the suggested locks. However, I'm still unsure of the necessity of locking for simple accessors like `empty()`  or `space()` that return integral types.

I was reading [this answer](https://stackoverflow.com/a/2252478) on StackOverflow which seems to suggest it might not be needed, but I'm not very well versed with multi-threaded programming so I still want more discussion around it.