@iProdigy 2022-06-07
the safest approach is to lock or privatize

in the public + no lock scenario,
1. to avoid undefined behavior with a concurrent read/write, you'd want the int to be updated atomically. since `std::atomic` isn't explicitly being used, the atomicity depends on the underlying architecture (and I would prefer the variable marked as `volatile` to avoid compiler optimizations that could get in the way of atomicity, but that's not possible here)
2. further, possible memory barriers could make concurrent reads on different threads yield different values, even if there is no concurrent write (though this degree of correctness probably isn't strictly required for this application)