@dnsge 2022-06-09
I added two more accessors, `first` and `last` which are the `get` equivalent of `front` and `back`. Adding the two made me wonder whether we should even bother supporting unchecked access of elements because it seems hard for callers to make guarantees about the size/emptiness of the queue.

We could just have `front`, `back`, and `at` which all return `boost::optional<T>` instead of having two for each type of access.