@leon-richardt 2022-06-09
> Adding the two made me wonder whether we should even bother supporting unchecked access of elements because it seems hard for callers to make guarantees about the size/emptiness of the queue.

Please correct me if I'm wrong but in strictly single-threaded use, callers should be able to make such guarantees, right?