@iProdigy 2022-06-09
> Please correct me if I'm wrong but in strictly single-threaded use, callers should be able to make such guarantees, right?

that's correct (no issues if all reads/writes done from the same thread)